Controls – create container view to store keyboard logic
This way I can keep the logic separate from the Joystick and they don't have to live on Settings, which now serves a different role.
